proto: move DnssecDnsHandle into dnssec module
