+++ crates/client/src/client/dnssec_client.rs
@@ -12,12 +12,12 @@ use std::sync::Arc;
 use futures_util::stream::Stream;
 
 use crate::client::Client;
+use crate::proto::dnssec::DnssecDnsHandle;
 use crate::proto::dnssec::TrustAnchor;
 use crate::proto::runtime::TokioTime;
 use crate::proto::xfer::{
     DnsExchangeBackground, DnsHandle, DnsRequest, DnsRequestSender, DnsResponse,
 };
-use crate::proto::DnssecDnsHandle;
 use crate::proto::ProtoError;
 
 /// A DNSSEC Client implemented over futures-rs.
