+++ crates/proto/src/dnssec/dnssec_dns_handle/nsec3_validation.rs
