+++ tests/integration-tests/tests/integration/dnssec_client_handle_tests.rs
@@ -7,7 +7,7 @@ use futures::executor::block_on;
 use tokio::runtime::Runtime;
 
 use hickory_client::client::{Client, ClientHandle, MemoizeClientHandle};
-use hickory_proto::dnssec::{Proof, TrustAnchor};
+use hickory_proto::dnssec::{DnssecDnsHandle, Proof, TrustAnchor};
 use hickory_proto::op::ResponseCode;
 use hickory_proto::rr::rdata::A;
 use hickory_proto::rr::Name;
@@ -15,7 +15,6 @@ use hickory_proto::rr::{DNSClass, RData, RecordType};
 use hickory_proto::runtime::TokioRuntimeProvider;
 use hickory_proto::tcp::TcpClientStream;
 use hickory_proto::udp::UdpClientStream;
-use hickory_proto::DnssecDnsHandle;
 use hickory_server::authority::{Authority, Catalog};
 
 use hickory_integration::example_authority::create_secure_example;
